Make regex into a multiple option
The (book/chapter) headings now contain # and |. These are special
characters in regexes and must be escpaed. Because it is tedious to
specify escaped regexes, we escape them in the tool. But we also want
to allow for multiple regexes, so we make the regex into a multiple
option.

To specify multiple regexes do:
python -m adhtools.split_text text.txt --regex '### |' --regex '### ||'