Extract all info (words, headers, metadata) in one go
Extracting the headers separately made the script crash on big xml files
(i.e. fiqh 1266), because it first has to loop through all the words
(without discarding information) before it gets to the markers.
Extracting all info in one go is more memory efficient, because
intermediate information is discarded when it is no longer necessary.