Fix adding of metadata
The metadata for books was not added to the chapter xml files, because
the element that was returned was cleared shortly afterwards. This
problem was fixed by making a copy of the metadata element.

The volume/chapter level metadata was concatenated (the metadata of
chapter 1 was concatenated to the metadata of chapter 2, etc.). This
was also fixed by making a copy of the metadata to which the chapter-
level metadata was concatenated.