Shorten string values that are too long for blacklab
We make sure the string values aren't to long, because
BlackLab doesn't allow values that are to long in dropdowns.
The default value of 94 was set emperically. It seems the
lengths of strings are caluclated differently in Java (the
max length in Java is 256).

Shortened strings get an X in front, to mark shortened strings.